添加对Restful的返回结果进行处理的方法，当未成功时，默认抛出运行时异常

+++ eladmin-common/src/main/java/com/lwohvye/utils/result/IResultInfo.java
@@ -26,7 +26,7 @@ import java.util.List;
  * @date 2021/1/7 17:26
  */
 public interface IResultInfo<T> extends Serializable {
-    long getBusinessCode();
+    int getBusinessCode();
 
     List<T> getContent();
